--- notebooks/conda_env.yml
@@ -1,22 +1,0 @@
-# installation from this file in conda: conda env create -f conda_env.yml
-# remove if already there: conda env remove -n jads
-name: jads
-channels:
-  - conda-forge
-dependencies:
-  - python=3.8
-  - pandas
-  - pyarrow
-  - jupyterlab
-  - jupyter_contrib_nbextensions
-  - ipykernel
-  - rise
-  - nb_black
-  - seaborn
-  - plotly
-  - altair
-  - statsmodels
-  - scikit-learn
-  - sktime
-# administer kernel: conda activate jads
-# python -m ipykernel install --name jads
